requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg[binary]>=3.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
Pillow>=10.0.0
stripe>=7.6.0
PyJWT>=2.8.0
